backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
fastapi==0.115.3
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
